+++ CHANGELOG.rst
@@ -1,7 +1,15 @@
 Changelog
 =========
 
-0.1.0 - unreleased
--------------------
+0.2.0 - 2026-05-07
+------------------
+
+* Added ``MARKDOWN_MIDDLEWARE_CONVERT_OPTIONS`` setting to pass conversion options
+  directly to ``html-to-markdown``'s ``ConversionOptions``. This allows fine-grained
+  control over the HTML-to-Markdown conversion.
+
+
+0.1.0 - 2026-05-06
+------------------
 
 * Initial release of ``django-markdown-middleware``